first shot at refactoring ci testing
