+++ pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+log_level = INFO
+minversion = 6.0
+required_plugins = pytest-xdist
+python_functions = test_*
+python_files = *.py
+testpaths =
+    tests
+markers =
+    slow: flag test as slow (deselect with '-m "not slow"')
+    online: flag test as requiring internet to work (deselect with '-m "not online"')
