+++ tests/commands/xinfo.py
@@ -1,0 +1,20 @@
+"""
+xinfo command test module
+"""
+
+
+from tests.utils import GefUnitTestGeneric, gdb_run_cmd, gdb_start_silent_cmd
+
+
+class XinfoCommand(GefUnitTestGeneric):
+    """`xinfo` command test module"""
+
+
+    def test_cmd_xinfo(self):
+        self.assertFailIfInactiveSession(gdb_run_cmd("xinfo $sp"))
+        res = gdb_start_silent_cmd("xinfo")
+        self.assertIn("At least one valid address must be specified", res)
+
+        res = gdb_start_silent_cmd("xinfo $sp")
+        self.assertNoException(res)
+        self.assertTrue(len(res.splitlines()) >= 7)
