+++ tests/utils.py
@@ -1,24 +1,76 @@
+"""
+Utility functions for testing
+"""
+
 import os
+import pathlib
 import platform
 import re
 import subprocess
-import sys
 import tempfile
-from pathlib import Path
-from typing import Iterable, Union, NewType, List
+import unittest
+import warnings
 
-TMPDIR = Path(tempfile.gettempdir())
-DEFAULT_TARGET = TMPDIR / "default.out"
-GEF_PATH = Path(os.getenv("GEF_PATH", "gef.py"))
-STRIP_ANSI_DEFAULT = True
-DEFAULT_CONTEXT = "-code -stack"
+from typing import Iterable, Union, List
+
+TMPDIR = pathlib.Path(tempfile.gettempdir())
 ARCH = (os.getenv("GEF_CI_ARCH") or platform.machine()).lower()
+BIN_LS = pathlib.Path("/bin/ls")
+BIN_SH = pathlib.Path("/bin/sh")
 CI_VALID_ARCHITECTURES_32B = ("i686", "armv7l")
 CI_VALID_ARCHITECTURES_64B = ("x86_64", "aarch64", "mips64el", "ppc64le")
 CI_VALID_ARCHITECTURES = CI_VALID_ARCHITECTURES_64B + CI_VALID_ARCHITECTURES_32B
 COVERAGE_DIR = os.getenv("COVERAGE_DIR", "")
+DEFAULT_CONTEXT = "-code -stack"
+DEFAULT_TARGET = TMPDIR / "default.out"
+GEF_DEFAULT_PROMPT = "gef➤  "
+GEF_DEFAULT_TEMPDIR = "/tmp/gef"
+GEF_PATH = pathlib.Path(os.getenv("GEF_PATH", "gef.py"))
+STRIP_ANSI_DEFAULT = True
 
-CommandType = NewType("CommandType", Union[str, Iterable[str]])
+
+CommandType = Union[str, Iterable[str]]
+
+
+class GdbAssertionError(AssertionError):
+    pass
+
+
+class GefUnitTestGeneric(unittest.TestCase):
+    """Generic class for command testing, that defines all helpers"""
+
+    @staticmethod
+    def assertException(buf):
+        """Assert that GEF raised an Exception."""
+        if not ("Python Exception <" in buf
+                or "Traceback" in buf
+                or "'gdb.error'" in buf
+                or "Exception raised" in buf
+                or "failed to execute properly, reason:" in buf):
+            raise GdbAssertionError("GDB Exception expected, not raised")
+
+    @staticmethod
+    def assertNoException(buf):
+        """Assert that no Exception was raised from GEF."""
+        if ("Python Exception <" in buf
+                or "Traceback" in buf
+                or "'gdb.error'" in buf
+                or "Exception raised" in buf
+                or "failed to execute properly, reason:" in buf):
+            raise GdbAssertionError(f"Unexpected GDB Exception raised in {buf}")
+
+        if "is deprecated and will be removed in a feature release." in buf:
+            lines = [l for l in buf.splitlines()
+                     if "is deprecated and will be removed in a feature release." in l]
+            deprecated_api_names = set([x.split()[1] for x in lines])
+            warnings.warn(
+                UserWarning(f"Use of deprecated API(s): {', '.join(deprecated_api_names)}")
+            )
+
+    @staticmethod
+    def assertFailIfInactiveSession(buf):
+        if "No debugging session active" not in buf:
+            raise AssertionError("No debugging session inactive warning")
 
 
 def is_64b() -> bool:
@@ -37,12 +89,13 @@ def _add_command(commands: CommandType) -> List[str]:
 
 
 def gdb_run_cmd(cmd: CommandType, before: CommandType = (), after: CommandType = (),
-                target: Path = DEFAULT_TARGET, strip_ansi: bool = STRIP_ANSI_DEFAULT) -> str:
+                target: pathlib.Path = DEFAULT_TARGET,
+                strip_ansi: bool = STRIP_ANSI_DEFAULT) -> str:
     """Execute a command inside GDB. `before` and `after` are lists of commands to be executed
     before (resp. after) the command to test."""
     command = ["gdb", "-q", "-nx"]
     if COVERAGE_DIR:
-        coverage_file = Path(COVERAGE_DIR) / os.getenv("PYTEST_XDIST_WORKER")
+        coverage_file = pathlib.Path(COVERAGE_DIR) / os.getenv("PYTEST_XDIST_WORKER", "gw0")
         command += _add_command([
             "pi from coverage import Coverage",
             f"pi cov = Coverage(data_file=\"{coverage_file}\","
@@ -73,9 +126,9 @@ def gdb_run_cmd(cmd: CommandType, before: CommandType = (), after: CommandType =
     while not result:
         try:
             result = output.decode("utf-8")
-        except UnicodeDecodeError as e:
-            faulty_idx_start = int(e.start)
-            faulty_idx_end = int(e.end)
+        except UnicodeDecodeError as ude:
+            faulty_idx_start = int(ude.start)
+            faulty_idx_end = int(ude.end)
             output = output[:faulty_idx_start] + output[faulty_idx_end:]
 
     if strip_ansi:
@@ -85,7 +138,7 @@ def gdb_run_cmd(cmd: CommandType, before: CommandType = (), after: CommandType =
 
 
 def gdb_run_silent_cmd(cmd: CommandType, before: CommandType = (), after: CommandType = (),
-                       target: Path = DEFAULT_TARGET,
+                       target: pathlib.Path = DEFAULT_TARGET,
                        strip_ansi: bool = STRIP_ANSI_DEFAULT) -> str:
     """Disable the output and run entirely the `target` binary."""
     before = [*before, "gef config context.clear_screen False",
@@ -95,14 +148,14 @@ def gdb_run_silent_cmd(cmd: CommandType, before: CommandType = (), after: Comman
 
 
 def gdb_run_cmd_last_line(cmd: CommandType, before: CommandType = (), after: CommandType = (),
-                          target: Path = DEFAULT_TARGET,
+                          target: pathlib.Path = DEFAULT_TARGET,
                           strip_ansi: bool = STRIP_ANSI_DEFAULT) -> str:
     """Execute a command in GDB, and return only the last line of its output."""
     return gdb_run_cmd(cmd, before, after, target, strip_ansi).splitlines()[-1]
 
 
 def gdb_start_silent_cmd(cmd: CommandType, before: CommandType = (), after: CommandType = (),
-                         target: Path = DEFAULT_TARGET,
+                         target: pathlib.Path = DEFAULT_TARGET,
                          strip_ansi: bool = STRIP_ANSI_DEFAULT,
                          context: str = DEFAULT_CONTEXT) -> str:
     """Execute a command in GDB by starting an execution context. This command
@@ -116,52 +169,28 @@ def gdb_start_silent_cmd(cmd: CommandType, before: CommandType = (), after: Comm
 
 def gdb_start_silent_cmd_last_line(cmd: CommandType, before: CommandType = (),
                                    after: CommandType = (),
-                                   target: Path = DEFAULT_TARGET,
+                                   target: pathlib.Path = DEFAULT_TARGET,
                                    strip_ansi=STRIP_ANSI_DEFAULT) -> str:
     """Execute `gdb_start_silent_cmd()` and return only the last line of its output."""
     return gdb_start_silent_cmd(cmd, before, after, target, strip_ansi).splitlines()[-1]
 
 
 def gdb_test_python_method(meth: str, before: str = "", after: str = "",
-                           target: Path = DEFAULT_TARGET,
+                           target: pathlib.Path = DEFAULT_TARGET,
                            strip_ansi: bool = STRIP_ANSI_DEFAULT) -> str:
     brk = before + ";" if before else ""
     cmd = f"pi {brk}print({meth});{after}"
     return gdb_start_silent_cmd(cmd, target=target, strip_ansi=strip_ansi)
 
 
-def include_for_architectures(valid_architectures: Iterable[str] = CI_VALID_ARCHITECTURES):
-    def wrapper(f):
-        def inner_f(*args, **kwargs):
-            if ARCH in valid_architectures:
-                f(*args, **kwargs)
-            else:
-                sys.stderr.write(f"SKIPPED for {ARCH}  ")
-                sys.stderr.flush()
-        return inner_f
-    return wrapper
-
-
-def exclude_for_architectures(invalid_architectures: Iterable[str] = ()):
-    def wrapper(f):
-        def inner_f(*args, **kwargs):
-            if ARCH not in invalid_architectures:
-                f(*args, **kwargs)
-            else:
-                sys.stderr.write(f"SKIPPED for {ARCH}  ")
-                sys.stderr.flush()
-        return inner_f
-    return wrapper
-
-
-def _target(name: str, extension: str = ".out") -> Path:
+def _target(name: str, extension: str = ".out") -> pathlib.Path:
     target = TMPDIR / f"{name}{extension}"
     if not target.exists():
         raise FileNotFoundError(f"Could not find file '{target}'")
     return target
 
 
-def start_gdbserver(exe: Union[str, Path] = _target("default"),
+def start_gdbserver(exe: Union[str, pathlib.Path] = _target("default"),
                     port: int = 1234) -> subprocess.Popen:
     """Start a gdbserver on the target binary.
 
@@ -186,4 +215,3 @@ def stop_gdbserver(gdbserver: subprocess.Popen) -> None:
     if gdbserver.poll() is None:
         gdbserver.kill()
         gdbserver.wait()
-    return
